Refactor game loop into systems module
