+++ src/game/systems/index.js
@@ -1,1 +1,2 @@
 export { CONFIG, createGameState, resetGameState } from "./state.js";
+export { initializeGameLoop, startGame, handleCanvasClick } from "./loop.js";
